backend/requirements.txt
# ===========================================
# Cursor-Writing 后端依赖
# ===========================================

# ----- 核心框架 -----
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ----- 异步支持 -----
aiofiles>=23.2.0
websockets>=12.0
python-multipart>=0.0.6

# ----- LLM 客户端 -----
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.27.0

# ----- 数据处理 -----
pyyaml>=6.0.0
python-dotenv>=1.0.0

# ----- 工具 (预留：爬虫/搜索) -----
# duckduckgo-search>=8.1.1
# beautifulsoup4>=4.12.0
# lxml>=5.3.0
# requests>=2.31.0

# ----- 开发/测试 -----
pytest>=8.0.0
pytest-asyncio>=0.23.0
